% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{CopulaModel_1stage}
\alias{CopulaModel_1stage}
\title{Title of this function (do later, function still temporary)}
\usage{
CopulaModel_1stage(data, time, status, cluster, covariates, init.values,
  marginal = "Weibull", copula = "Clayton", n.piecewise = 20)
}
\arguments{
\item{data}{Input dataframe containing all variables.}

\item{time}{Which variable name is the time covariate?}

\item{status}{Which variable name is the status covariate? (more explanation?)}

\item{cluster}{Which variable name is the cluster covariate?}

\item{covariates}{Vector of 1 or more other covariates to be included in the model. (NOTE: MULTIPLE COVARIATES NEED TO BE TESTED!!!)}

\item{init.values}{Initial values for parameters: 
\itemize{
\item Clayton & Weibull: \deqn{\lambda, \rho, \theta and \beta's}
\item Clayton & Piecewise Exponential: \deqn{\lambda's, \theta and \beta's} Note that the number of lambda's is equal to the number of pieces of the piecewise exponential.
}}

\item{marginal}{Which marginal to use? Can either be \code{"Weibull"} or \code{"PiecewiseExp"} for Piecewise Exponential.}

\item{copula}{Which copula to user? Can either be \code{"Clayton"} or \code{"GH"} for Gumbel-Hougaard.}

\item{n.piecewise}{For \code{marginal="PiecewiseExp"}, how many pieces should the Piecewise Exponential have? (Default = 20)}
}
\value{
Not yet final, so far: list object with: parameters (est & se) + kendal + cov matrix. Add more information?(NOTE: EXAMPLES CAN LATER BE RUN)
}
\description{
Description of this function. (Note: you can also add other fields like references, authors,... See which ones are necessary)
}
\examples{
\dontrun{
data("insem",package="Sunclarco")
result <- CopulaModel_1stage(data=insem,time="Time",status="Status",cluster="Herd",covariates="Heifer",
		init.values=c(log(0.05),log(0.5),log(0.5),-0.05),
		marginal="Weibull",copula="Clayton")
result
}
}

