% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sunclarco.R
\docType{package}
\name{Sunclarco}
\alias{Sunclarco}
\alias{Sunclarco-package}
\title{Survival Analysis using Copula's (TITLE)}
\description{
Survival analysis for unbalanced clusters using archimedes copula's.
This is the description of the package, add as many lines as necessary.
}
\author{
Leen Prenen

Ewoud De Troyer

Roel Braekers

Luc Duchateau
}
\references{
Reference 1

Reference 2
}

