% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{insem}
\alias{insem}
\title{Insemination Data}
\format{A dataframe with 10513 rows and 5 columns.}
\description{
Description of the data.
Can have multiple lines. (Note that other fields like references, authors,... can be added)
}

