% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SunclarcoModel}
\alias{SunclarcoModel}
\title{Sunclarco Model}
\usage{
SunclarcoModel(data, time, status, clusters, covariates, stage = 1,
  copula = "Clayton", marginal = "Weibull", n.piecewise = 20,
  init.values = NULL, verbose = FALSE, summary.print = TRUE)
}
\arguments{
\item{data}{Input dataframe containing all variables.}

\item{time}{Which variable name is the time covariate?}

\item{status}{Which variable name is the status covariate? (more explanation?)}

\item{clusters}{Which variable name is the cluster covariate?}

\item{covariates}{Vector of 1 or more other covariates to be included in the model. If necessary a covariate should be a factor in the data frame. (NOTE: MULTIPLE COVARIATES NEED TO BE TESTED!!!)}

\item{stage}{Can be 1 or 2 for 1 or 2 staged approached. See Details for more information.}

\item{copula}{Which copula to user? Can either be \code{"Clayton"} or \code{"GH"} for Gumbel-Hougaard.}

\item{marginal}{Which marginal to use? Can either be \code{"Weibull"}, \code{"PiecewiseExp"} for Piecewise Exponential or \code{"Cox"} for Cox Regression.}

\item{n.piecewise}{For \code{marginal="PiecewiseExp"}, how many pieces should the Piecewise Exponential have? (Default = 20)}

\item{init.values}{Initial values for parameters. This depends on the choice of the parameters \code{stage}, \code{copula} and \code{marginal}. See Details for more information.
If no initial parameters are given, they will be chosen automatically (See Details for more information).}

\item{verbose}{Print some in-between results as well as computation progress.}

\item{summary.print}{Logical value to print a short summary at the end of the computation.}
}
\value{
S3 List object
\item{Parameters}{Data frame containing estimates and standard errors of parameters.}
\item{Kendall_Tau}{Vector containing estimate and standard error of Kendall's Tau.}
\item{ParametersCov}{If available, covariance matrix of the parameters. For 2-stage approaches this is only available for the Weibull marginal.}
\item{logllh}{The log-likelihood value.}
\item{parameter.call}{A list containing all arguments given to the function, as well as the initial parameter values and the elapsed time.}
}
\description{
Model for Survival Analysis of Unbalanced Clusters using Archimedes Copula's
 (Note: you can also add other fields like references, authors,... See which ones are necessary)
}
\details{
WRITE DIFFERENT SECTIONS
SECTION 1: Which marginals + copula's are available for stage 1 and 2
SECTION 2: Which parameters have to be given initially in all scenarios (stage 1 and 2) and in which order!!
(note: the order will always be like this:  lambdas, rho, theta, betas)
SECTION 3: How are the initial parameters chosen if done automatically?
}
\examples{
\dontrun{
data("insem",package="Sunclarco")
result1 <- SunclarcoModel(data=insem,time="Time",status="Status",
                          clusters="Herd",covariates="Heifer",
                          stage=1,copula="Clayton",marginal="Weibull")

summary(result1)

result2 <- SunclarcoModel(data=insem,time="Time",status="Status",
                          clusters="Herd",covariates="Heifer",
                          stage=2,copula="Clayton",marginal="Cox",
                          verbose=TRUE)
summary(result2)
## ADD MORE EXAMPLES? => THESE EXAMPLES SHOULD BE USED IN VIGNETTE?
}
}

